fix the cost page not rendering when Kubecost is not connected, and add error message to let clients know they need to run Kubecost to see cluster data
